+++ domain/posts.go
@@ -3,7 +3,7 @@ package domain
 import "time"
 
 type Posts struct {
-	ID        int       `json:"id" bson:"id" gorm:"primaryKey;autoIncrement"`
+	ID        int64     `json:"id" bson:"id" gorm:"primaryKey;autoIncrement"`
 	UserID    string    `json:"user_id" bson:"user_id"`
 	Content   string    `json:"content" bson:"content"`
 	ImageUrl  string    `json:"image_url" bson:"image_url"`
